+++ addon/adapters/offline.js
@@ -429,6 +429,101 @@ export default DS.Adapter.extend({
   },
 
   /**
+    Performing batch update operaion.
+
+    @method batchUpdate
+    @param {DS.Store} store
+  	@param {DS.Model[]|DS.Model} models Is array of models or single model for batch update.
+    @return {Promise}
+  */
+  batchUpdate(store, models) {
+    if (Ember.isEmpty(models)) {
+      return Ember.RSVP.resolve(models);
+    }
+
+    models = Ember.isArray(models) ? models : Ember.A([models]);
+
+    const dexieService = this.get('dexieService');
+    const db = dexieService.dexie(this.get('dbName'), store);
+
+    const modelsMap = {};
+
+    models.forEach((model) => {
+      const snapshot = model._createSnapshot();
+      const modelName = snapshot.modelName;
+      const serializer = store.serializerFor(modelName);
+
+      modelsMap[modelName] = modelsMap[modelName] || { add: [], update: [], delete: [] };
+
+      const dirtyType = model.get('dirtyType') || model.hasChangedBelongsTo() ? 'updated' : undefined;
+
+      switch (dirtyType) {
+        case 'created':
+          modelsMap[modelName].add.push(serializer.serialize(snapshot, { includeId: true }));
+          break;
+
+        case 'updated':
+          modelsMap[modelName].update.push(serializer.serialize(snapshot, { includeId: true }));
+          break;
+
+        case 'deleted':
+          modelsMap[modelName].delete.push(model.get('id'));
+          break;
+      }
+    });
+
+    const tableNames = Object.keys(modelsMap);
+    const batchUpdateOperation = (db) => db.transaction('rw', tableNames, () => new Dexie.Promise((resolve, reject) => {
+      const promises = [];
+      tableNames.forEach((tableName) => {
+        const table = db.table(tableName);
+        const mapItem = modelsMap[tableName];
+
+        if (mapItem.add.length) {
+          promises.push(table.bulkAdd(mapItem.add));
+        }
+
+        if (mapItem.update.length) {
+          promises.push(table.bulkPut(mapItem.update));
+        }
+
+        if (mapItem.delete.length) {
+          promises.push(table.bulkDelete(mapItem.delete));
+        }
+      });
+
+      Dexie.Promise.all(promises).then(() => {
+        Dexie.Promise.all(models.map((model) => {
+          const modelName = model.constructor.modelName;
+          const { id, dirtyType } = model.getProperties('id', 'dirtyType');
+
+          return dirtyType === 'deleted' ? Dexie.Promise.resolve(null) : db.table(modelName).get(id);
+        })).then((rawModels) => {
+          resolve(models.map((model, index) => {
+            Ember.run(store, store.unloadRecord, model);
+
+            const rawModel = rawModels[index];
+            if (rawModel) {
+              const modelName = model.constructor.modelName;
+              const modelClass = store.modelFor(modelName);
+              const serializer = store.serializerFor(modelName);
+
+              Ember.run(store, store.push, serializer.normalize(modelClass, rawModel));
+              Ember.run(model, model.rollbackAttributes);
+
+              return model;
+            }
+
+            return null;
+          }));
+        }).catch(reject);
+      }).catch(reject);
+    }));
+
+    return dexieService.performOperation(db, batchUpdateOperation);
+  },
+
+  /**
     Stores hash for performing bulk operaion into map.
 
     @method _storeHashForBulkOperation